+++ models/veiculo/sppo_veiculo_dia.sql
@@ -1,3 +1,4 @@
+-- depends_on: {{ ref('sppo_licenciamento_stu_staging') }}
 {{
     config(
         materialized="incremental",
@@ -6,6 +7,10 @@
         incremental_strategy="insert_overwrite",
     )
 }}
+
+{% if execute %}
+  {% set licenciamento_date = run_query(get_license_date()).columns[0].values()[0] %}
+{% endif %}
 
 WITH
   licenciamento AS (
@@ -25,20 +30,7 @@ WITH
   FROM
     {{ ref("sppo_licenciamento") }} --`rj-smtr`.`veiculo`.`sppo_licenciamento`
   WHERE
-  {%- if var("stu_data_versao") != "" %}
-    data = DATE("{{ var('stu_data_versao') }}")
-  -- Versão fixa do STU em 2024-03-25 devido à falha de atualização na fonte da dados (SIURB)
-  {%- elif var("run_date") >= "2024-03-01" and var("run_date") < "2024-03-16" %}
-    data = "2024-03-25"
-  -- Versão fixa do STU em 2024-04-09 para mar/Q2 devido à falha de atualização na fonte da dados (SIURB)
-  {%- elif var("run_date") >= "2024-03-16" %}
-    data = "2024-04-09"
-  {% else -%}
-    {%- if execute %}
-        {% set licenciamento_date = run_query("SELECT MIN(data) FROM " ~ ref("sppo_licenciamento") ~ " WHERE data >= DATE_ADD(DATE('" ~ var("run_date") ~ "'), INTERVAL 5 DAY)").columns[0].values()[0] %}
-    {% endif -%}
     data = DATE("{{ licenciamento_date }}")
-  {% endif -%}  
   ),
   gps AS (
   SELECT